# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HIX-AI-2API",
        "title": "HIX-AI-2API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/chat/completions": {
            "post": {
                "description": "OpenAI对话接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAI"
                ],
                "parameters": [
                    {
                        "description": "OpenAI对话请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OpenAIChatCompletionRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization API-KEY",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "model.OpenAIChatCompletionRequest": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OpenAIChatMessage"
                    }
                },
                "model": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                }
            }
        },
        "model.OpenAIChatMessage": {
            "type": "object",
            "properties": {
                "content": {},
                "role": {
                    "type": "string"
                }
            }
        }
    }
}